PHP Enum Class
This is a simple class with only TWO public functions.  These are PUT and GET.  PUT will put a new enum into the array and GET will get an enum from the array.  The array is private so no one can modify it.  The program checks to see if the enum has already been declared and if it has - it can not be modified.  Several methods of both putting new enums in and getting enums out are provided.  You can use an array, a list, and a single enum.  Unlike with other languages though, all enums MUST be enclosed in single or double quotes.  This is a problem with PHP.  Only DEFINEd values can be used without single or double quotes.  But as a simple way to have enums - this should work without any problems.  The code is heavily commented.  Feel free to modify the comments or change them if you feel you have a better way of doing so.