Added in the new __get() and __set() functions
With the addition of the __get() and __set() functions - now you can truly make this look like enums.  Example:

$c->RED_LIGHT = 85;

The above would call the __set() function and create the RED_LIGHT entry into the enum array.  Its value would be set as 85.  The
new variable $last_value would be set to 86.  One more than the last value given.  New variables that do not have a value assigned
to them would have this new value assigned to them and the $last_value would then be incremented as needed.